+++ polymer_pantheon_drupal.poly_info.yml
@@ -1,0 +1,1 @@
+name: Polymer Pantheon Drupal
