+++ src/Plugin/Template/DrushSiteYaml.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace DigitalPolygon\PolymerPantheon\Drupal\Polymer\Plugin\Template;
+namespace DigitalPolygon\Polymer\polymer_pantheon_drupal\Plugin\Template;
 
-use DigitalPolygon\Polymer\Robo\Template\Template;
-use DigitalPolygon\Polymer\Robo\Template\Token;
+use DigitalPolygon\Polymer\Core\Robo\Template\Template;
+use DigitalPolygon\Polymer\Core\Robo\Template\Token;
 
 class DrushSiteYaml extends Template
 {
